Treat retryable error as transient error during ami creation or instance start/stop.
The current scenario is, when there is a timeout error or connection reset error the build is allowed to continue. However when there is a rate limit error the build is aborted.

The justification for the fix is, whenever an ami is created or instance is started or stopped the status is checked every 2 seconds for 5 minutes (default AWS_TIMEOUT_SECONDS). If one of the status check fails it goes into exponential backoff retry loop in aws go sdk. The retry limit is set to only 11. If AWS is taking time   maximum 150 calls are made to AWS, whereas if there are 11 successive rate limit errors while in retry loop, for the build, it is aborted it in the middle.